--- src/components/Home.jsx
@@ -1,14 +1,0 @@
-import LayoutWrapper from './LayoutWrapper';
-import TypingField from './TypingField';
-import HomepageHeader from './HomepageHeader';
-
-const Home = () => {
-  return (
-    <LayoutWrapper>
-      <HomepageHeader />
-      <TypingField />
-    </LayoutWrapper>
-  );
-};
-
-export default Home;
