+++ src/routes/Home.jsx
@@ -1,0 +1,14 @@
+import LayoutWrapper from '../components/LayoutWrapper';
+import TypingField from '../components/TypingField';
+import HomepageHeader from '../components/HomepageHeader';
+
+const Home = () => {
+  return (
+    <LayoutWrapper>
+      <HomepageHeader />
+      <TypingField />
+    </LayoutWrapper>
+  );
+};
+
+export default Home;
